+++ src/Utils/OffsetManager.cpp
@@ -22,7 +22,8 @@ unsigned int OffsetManager::offsetForType(PatchType type)
             GEODE_WINDOWS(return 0x607008);
             GEODE_ANDROID32(return 0x4740f0 - 0x10000);
             GEODE_ANDROID64(return 0x9473c0 - 0x100000);
-            GEODE_INTEL_MAC(return );
+            GEODE_INTEL_MAC(return 0x7ff4d8);
+            GEODE_ARM_MAC(return );
             GEODE_IOS(return );
     
             #endif
@@ -39,6 +40,7 @@ unsigned int OffsetManager::offsetForType(PatchType type)
             GEODE_ANDROID32(return 0x46ce38 - 0x10000);
             GEODE_ANDROID64(return 0x9384b8 - 0x100000);
             GEODE_INTEL_MAC(return 0x823b00);
+            GEODE_ARM_MAC(return );
             GEODE_IOS(return 0x642b60);
     
             #endif